1546번 평균 int -> double로 변경

+++ backJoon_java/src/code/backJoon_1546.java
@@ -6,9 +6,9 @@ public class backJoon_1546 {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		int n = scan.nextInt();
-		int score = 0;
-		int max = 0;
-		int ave = 0;
+		double score = 0;
+		double max = 0;
+		double ave = 0;
 		for(int i=0; i<n; i++){
 			score=scan.nextInt();
 			ave += score; 
@@ -16,7 +16,7 @@ public class backJoon_1546 {
 				max = score;
 			}
 		}
-		System.out.println((double)ave/(double)max*100/n);
+		System.out.println(ave/max*100/n);
 		scan.close();
 	}
 }
